update: Saner rounding of sector building money and work
buildeff() rounds work and money up.  Until recently, fractions could
only occur on tear-down, but with customized costs they can now also
occur on build-up.

The previous commit changed unit building to round money and work
randomly.  Before, money was rounded down, and work was rounded up.

Round them randomly for sectors as well, for consistency.

+++ include/prototypes.h
@@ -712,7 +712,7 @@ extern int update_removewants(void);
 extern void revolt(struct sctstr *);
 extern void guerrilla(struct sctstr *);
 /* sect.c */
-extern int buildeff(struct sctstr *);
+extern double buildeff(struct sctstr *);
 extern void do_fallout(struct sctstr *, int);
 extern void spread_fallout(struct sctstr *, int);
 extern void decay_fallout(struct sctstr *, int);
